+++ crates/bevy_cef_core/src/lib.rs
@@ -6,6 +6,10 @@ mod render_process;
 mod util;
 
 pub mod prelude {
+    #[cfg(target_os = "windows")]
+    pub use crate::browser_process::cef_command::{BrowsersProxy, CefCommand};
+    #[cfg(target_os = "windows")]
+    pub use crate::browser_process::cef_thread::{drain_commands, init_cef_browsers};
     pub use crate::browser_process::*;
     #[cfg(target_os = "macos")]
     pub use crate::debug::*;